Extract EditCommandParser from Parser class
Parser class implements all different
command parsers as private methods within the Parser class.
Parser::parseCommand is responsible for performing an initial
parsing on the given input string and from there determine
the correct method to parse the String to.

This causes the Parser file to become very long
containing all command parser methods
and their associated helper methods that is only utilized in that
one method itself.

The design can be improved by applying the Single Responsibility
Principle as follows: limit the Parser's responsibility to determining
which type of command is being parsed and move the responsibility
of parsing each type of command to a separate *CommandParser class.

As a step toward the design proposed above, let's move the functionality
of parsing of the edit command from Parser to a new
EditCommandParser class.
